Working on authentication and permissions

+++ src/contexts/EmployeeContext.jsx
@@ -24,15 +24,23 @@ export const EmployeeProvider = ({ children }) => {
         observations: ""
     });
     const [area, setArea] = useState("");
+    const [token, setToken] = useState(localStorage.getItem("token") || null);
+    const [userRole, setUserRole] = useState(null);
+    const [picture, setPicture] = useState(null);
+    const [cv, setCV] = useState(null);
 
     useEffect (() =>{
-        fetch("http://localhost:3000/profiles")
-        .then(resp => resp.json())
-        .then(info => {
-          setEmployees(info);
-        })
-        .catch(err => console.log("Error fetching employees in front: ", err));
-    },[refresh]);
+        if(token){
+            fetch("http://localhost:3000/profiles", {
+                headers: {"Authorization": `Bearer ${token}`}
+            })
+            .then(resp => resp.json())
+            .then(info => {
+            setEmployees(info);
+            })
+            .catch(err => console.log("Error fetching employees in front: ", err));
+        }
+    },[refresh, token, userRole]);
 
     const onSearchInfo = event =>{
         setSearchField(event.target.value);
@@ -92,17 +100,28 @@ export const EmployeeProvider = ({ children }) => {
         })
         .then(resp => resp.json())
         .then(data =>{
-            if(data != "success"){
+            if(!data.token){     // data != "success"){
                 throw new Error ("WRONG CREDENTIALS");
+            } else {
+                localStorage.setItem("token", data.token);
+                setToken(data.token);
+                
+                const decoded = JSON.parse(atob(data.token.split(".")[1]));
+                setUserRole(decoded.role);
+                
+
             }
             navigate("/employees");
 
-            // if(data === "success"){
-            //     navigate("/employees");
-            // } 
         }).catch(err =>{
             setErrorMessage(err.message);
         })
+    }
+
+    const onExit = () =>{
+        localStorage.removeItem("token");
+        setToken(null);
+        setUserRole(null);
     }
 
     const onCandidateSave = (navigate, newCandidateData) =>{
@@ -134,14 +153,18 @@ export const EmployeeProvider = ({ children }) => {
     }
 
     const deleteEmployee = employee =>{
-        fetch(`http://localhost:3000/profile/${employee.id}`,{
-            method: 'delete',
-            headers: {'Content-Type': 'application/json'},
-            body: JSON.stringify(employee)
-        }).then( resp =>{
-            setRefresh(prev => !prev);
-            resp.json();
-        })
+        if(userRole === "admin"){
+            fetch(`http://localhost:3000/profile/${employee.id}`,{
+                method: 'delete',
+                headers: {'Authorization': `Bearer ${token}`},
+                body: JSON.stringify(employee)
+            }).then( resp =>{
+                setRefresh(prev => !prev);
+                resp.json();
+            }).catch(err => {
+                setErrorMessage(err.message);
+            })
+        }
     }
 
     return (
@@ -153,6 +176,8 @@ export const EmployeeProvider = ({ children }) => {
                 errorMessage,
                 filteredStaff,
                 area,
+                token,
+                userRole,
                 onEmployeeClick,
                 setSelectedEmployee,
                 updateEmployee,
@@ -164,6 +189,9 @@ export const EmployeeProvider = ({ children }) => {
                 deleteEmployee,
                 onSearchInfo,
                 onAreaChange,
+                onExit,
+                setCV,
+                setPicture
             }}>
             {children}
         </EmployeeContext.Provider>
